Move withActionAfterWait to matching helper
